+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - exception thrown when `!history` builtin method was used with invalid argument
 - correctly wrap lists (such as `ls` or `dir` listing) at the longest word
+- interpretation of SHV type hints and inclusion of `IS_UPDATABLE` flag in
+  method parameter validation
 
 
 ## [0.8.1] - 2025-11-26